add distro method to declare modules as required
Currently in the non-monolithic build modules are compiled into the base module when their interface file contains a xml element required with the attribute val and the value true. This setting is determined for all configurations and for all distributions.

This patch adds an options to configure a module as basemodule for a specific distribution (using the build.conf variable DISTRO) by defining the value val as a comma separated list of distributions, like <required val="debian" /> or <required val="debian,gentoo" />.